Continuous Placement
instead of click to spawn an Actor and next click to drop actor at chosen cell, make click + drag to continually spawn a selected actor and if Cell position is valid drop it on Grid.
Clean up of old classes, Actor components not required anymore as we are checking placement based on Cell Location in Grid.